Rename files to match convention
ConstantPropagationAnalysis.py -> constant_propagation_analysis
AbstractDFA.py -> abstract_dfa.py
DataFlowAnalysis.py -> data_flow_analysis.py
